backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
redis>=5.0.0
pyyaml>=6.0
requests>=2.31.0
openai>=1.3.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
pytz>=2023.0
circuitbreaker>=1.4.0
pyjwt[crypto]>=2.8.0
bcrypt==3.2.0
passlib[bcrypt]>=1.7.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
cryptography>=41.0.0
structlog>=23.0.0
loguru>=0.7.0
asyncer>=0.0.11
typing-extensions>=4.8.0
anyio>=3.7.0,<4.0.0
idna>=3.4
certifi>=2023.0
h11>=0.14.0
sniffio>=1.3.0
charset-normalizer>=3.3.0
urllib3>=2.0.0
annotated-types>=0.6.0
click>=8.1.0
colorama>=0.4.0
exceptiongroup>=1.1.0
hiredis>=2.0.0
itsdangerous>=2.1.0
jinja2>=3.1.0
markupsafe>=2.1.0
orjson>=3.9.0
python-dateutil>=2.8.0
six>=1.16.0
starlette>=0.27.0
tzdata>=2023.0
watchfiles>=0.21.0
websockets>=12.0
psutil>=5.9.0
GPUtil>=1.4.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.24.0
openwakeword>=0.5.0
duckduckgo-search>=6.0.0
aiohttp>=3.9.0
huggingface_hub>=0.19.0

python-multipart>=0.0.9

# Development and Quality Tools
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
isort>=5.12.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
